+++ cmd/contact.go
@@ -56,28 +56,7 @@ func init() {
 	contactCmd.AddCommand(contactViewCmd)
 	contactCmd.AddCommand(contactEmailCmd)
 
-	// Add
-	contactFlags(contactAddCmd)
-	contactAddCmd.Flags().BoolP("fast", "f", false, "create a contact entry only using a name and email")
-
-	// Edit
-	contactFlags(contactEditCmd)
-
-	// View
-	contactViewCmd.Flags().StringP("output", "o", "", "save the results to a file")
-	contactViewCmd.Flags().Bool("force", false, "force the results to be saved to an existing file")
-}
-
-func contactFlags(cmd *cobra.Command) {
-	cmd.Flags().StringP("name", "n", "", "name for this contact")
-	cmd.Flags().StringP("email", "e", "", "email for this contact")
-	cmd.Flags().String("phone", "", "phone number for this contact")
-
-	cmd.Flags().BoolP("contacted", "c", false, "contact was made")
-
-	cmd.Flags().Bool("lead", false, "contact is a lead")
-	cmd.Flags().Bool("customer", false, "contact is a customer")
-	cmd.Flags().Bool("subscriber", false, "contact is a subscriber")
-	cmd.Flags().Bool("advocate", false, "contact is a advocate")
-	cmd.Flags().String("other", "", "some other contact relationship")
+	// // View
+	// contactViewCmd.Flags().StringP("output", "o", "", "save the results to a file")
+	// contactViewCmd.Flags().Bool("force", false, "force the results to be saved to an existing file")
 }