+++ code/db/contact.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/gosimple/slug"
 	"github.com/jinzhu/gorm"
+	"github.com/nyaruka/phonenumbers"
 )
 
 // DB is accessable database.
@@ -21,21 +22,12 @@ type Contact struct {
 
 	Notes []Note `json:"notes,omitempty"`
 
-	Relationship   Relationship `json:"relationship"`
-	RelationshipID uint         `json:"-"`
-
-	Contacted bool `json:"contacted"`
-}
-
-// Relationship describes what this contact does.
-type Relationship struct {
-	gorm.Model
-
 	Lead       bool   `json:"lead,omitempty"`
 	Advocate   bool   `json:"advocate,omitempty"`
 	Customer   bool   `json:"customer,omitempty"`
 	Subscriber bool   `json:"subscriber,omitempty"`
 	Other      string `json:"other"`
+	Contacted  bool   `json:"contacted"`
 }
 
 // Create a contact in the database.
@@ -46,9 +38,16 @@ func (c *Contact) Create() error {
 		return errors.New("missing a name or email")
 	}
 
+	// Format the phone number
+	if c.Number != "" {
+		if number, err := phonenumbers.Parse(c.Number, "US"); err == nil {
+			c.Number = phonenumbers.Format(number, phonenumbers.NATIONAL)
+		}
+	}
+
 	// All subscribers are customers, but not all customers are subscribers.
-	if c.Relationship.Subscriber {
-		c.Relationship.Customer = true
+	if c.Subscriber {
+		c.Customer = true
 	}
 
 	c.Slug = slug.Make(c.Name)
@@ -57,14 +56,22 @@ func (c *Contact) Create() error {
 }
 
 // Update a contact that is in the database.
-func (c *Contact) Update(u Contact) error {
+func (c *Contact) Update() error {
 
 	// All subscribers are customers, but not all customers are subscribers.
-	if u.Relationship.Subscriber {
-		u.Relationship.Customer = true
+	if c.Subscriber {
+		c.Customer = true
 	}
 
-	return DB.Model(&c).Updates(&u).Error
+	// Format the phone number
+	if c.Number != "" {
+		if number, err := phonenumbers.Parse(c.Number, "US"); err == nil {
+			c.Number = phonenumbers.Format(number, phonenumbers.NATIONAL)
+		}
+	}
+
+	// return DB.Model(&c).Updates(&c).Error
+	return DB.Save(&c).Error
 }
 
 // Remove a contact that is in the database.
@@ -74,30 +81,7 @@ func (c *Contact) Remove() error {
 		return errors.New("need an ID")
 	}
 
-	// Remove relationship entry.
-	rel := Relationship{}
-	DB.Model(&c).Related(&rel)
-
-	if err := DB.Delete(&c).Error; err != nil {
-		return err
-	}
-
-	return DB.Delete(&rel).Error
-}
-
-// Fill the contact with all additional data.
-func (c *Contact) Fill() error {
-
-	var e Relationship
-
-	if err := DB.Model(&c).Related(&e).Error; err != nil {
-		return err
-	}
-
-	c.Relationship = e
-
-	return nil
-
+	return DB.Delete(&c).Error
 }
 
 // Query a contact given an ID.
@@ -107,11 +91,7 @@ func (c *Contact) Query() error {
 		return errors.New("need an ID")
 	}
 
-	if err := DB.First(&c).Error; err != nil {
-		return err
-	}
-
-	return c.Fill()
+	return DB.First(&c).Error
 
 }
 
@@ -129,9 +109,7 @@ func QueryContacts() ([]Contact, error) {
 // Search returns a single contact that matches c.
 func (c *Contact) Search() (err error) {
 
-	err = DB.Where(&c).First(&c).Error
-
-	return c.Fill()
+	return DB.Where(&c).First(&c).Error
 
 }
 
@@ -139,10 +117,6 @@ func (c *Contact) Search() (err error) {
 func (c *Contact) SearchMultiple() (result []Contact, err error) {
 
 	err = DB.Where(&c).Find(&result).Error
-
-	for y := range result {
-		err = result[y].Fill()
-	}
 
 	return
 }