+++ code/manage/contact-delete.go
@@ -1,8 +1,9 @@
 package manage
 
 import (
-	"broadcastle.co/code/crm/code/db"
+	"broadcastle.co/code/crm/code/tui"
 	"broadcastle.co/code/crm/code/utils"
+	"github.com/rivo/tview"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -12,8 +13,8 @@ func ContactRemove(cmd *cobra.Command, args []string) {
 
 	logrus.Debug("beginning the contact removal process")
 
-	db.Init()
-	defer db.Close()
+	Init()
+	defer Close()
 
 	contacts, err := utils.Contacts(cmd, args)
 	if err != nil {
@@ -22,11 +23,15 @@ func ContactRemove(cmd *cobra.Command, args []string) {
 
 	for x := range contacts {
 
-		if err := contacts[x].Remove(); err != nil {
-			logrus.Fatal(err)
+		app := &tui.App{
+			tview.NewApplication(),
 		}
 
-		logrus.Debugf("%s was removed as a contact", contacts[x].Name)
+		modal := app.Remove(contacts[x])
+
+		if err := app.SetRoot(modal, true).SetFocus(modal).Run(); err != nil {
+			logrus.Fatal(err)
+		}
 
 	}