+++ code/manage/contact-edit.go
@@ -1,8 +1,9 @@
 package manage
 
 import (
-	"broadcastle.co/code/crm/code/db"
+	"broadcastle.co/code/crm/code/tui"
 	"broadcastle.co/code/crm/code/utils"
+	"github.com/rivo/tview"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -10,8 +11,8 @@ import (
 // ContactEdit changes a contact based on the given ID.
 func ContactEdit(cmd *cobra.Command, args []string) {
 
-	db.Init()
-	defer db.Close()
+	Init()
+	defer Close()
 
 	contacts, err := utils.Contacts(cmd, args)
 	if err != nil {
@@ -20,62 +21,17 @@ func ContactEdit(cmd *cobra.Command, args []string) {
 
 	for _, contact := range contacts {
 
-		update, rel := db.Contact{}, db.Relationship{}
-
-		if err := db.DB.Model(&contact).Related(&rel).Error; err != nil {
-			logrus.Fatal(err)
+		app := &tui.App{
+			tview.NewApplication(),
 		}
 
-		update.Name, err = utils.CobraInput(cmd, "name", "Client Name", contact.Name, false, false)
-		if err != nil {
-			logrus.Fatal(err)
-		}
+		form := app.ContactForm(contact.ID)
 
-		update.Email, err = utils.CobraInput(cmd, "email", "Client Email", contact.Email, false, false)
-		if err != nil {
-			logrus.Fatal(err)
-		}
+		form.SetBorder(true).
+			SetTitle("Edit " + contact.Name).
+			SetTitleAlign(tview.AlignCenter)
 
-		update.Number, err = utils.CobraInput(cmd, "phone", "Client Phone Number", contact.Number, false, false)
-		if err != nil {
-			logrus.Fatal(err)
-		}
-
-		update.Contacted, err = utils.CobraInputBool(cmd, "contacted", "Contact was made", contact.Contacted, false)
-		if err != nil {
-			logrus.Warn(err)
-		}
-
-		rel.Lead, err = utils.CobraInputBool(cmd, "lead", "Contact is a lead", rel.Lead, false)
-		if err != nil {
-			logrus.Warn(err)
-		}
-
-		rel.Subscriber, err = utils.CobraInputBool(cmd, "subscriber", "Contact is a subscriber", rel.Subscriber, false)
-		if err != nil {
-			logrus.Warn(err)
-		}
-
-		if !rel.Subscriber {
-			rel.Customer, err = utils.CobraInputBool(cmd, "customer", "Contact is a customer", rel.Customer, false)
-			if err != nil {
-				logrus.Warn(err)
-			}
-		}
-
-		rel.Advocate, err = utils.CobraInputBool(cmd, "advocate", "Contact is a advocate", rel.Advocate, false)
-		if err != nil {
-			logrus.Warn(err)
-		}
-
-		rel.Other, err = utils.CobraInput(cmd, "other", "Other customer relationship", rel.Other, false, false)
-		if err != nil {
-			logrus.Warn(err)
-		}
-
-		update.Relationship = rel
-
-		if err := contact.Update(update); err != nil {
+		if err := app.SetRoot(form, true).SetFocus(form).Run(); err != nil {
 			logrus.Fatal(err)
 		}