+++ code/manage/contact-view.go
@@ -1,13 +1,10 @@
 package manage
 
 import (
-	"bytes"
-	"encoding/json"
-	"io/ioutil"
-	"os"
-
 	"broadcastle.co/code/crm/code/db"
+	"broadcastle.co/code/crm/code/tui"
 	"broadcastle.co/code/crm/code/utils"
+	"github.com/rivo/tview"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
@@ -15,57 +12,50 @@ import (
 // ContactView returns the selected contact given their ID's.
 func ContactView(cmd *cobra.Command, args []string) {
 
-	db.Init()
-	defer db.Close()
-
-	var result bytes.Buffer
+	Init()
+	defer Close()
 
 	contacts, err := utils.Contacts(cmd, args)
 	if err != nil {
 		logrus.Fatal(err)
 	}
 
-	if len(args) < 1 {
-		var err error
-
+	if len(contacts) < 1 {
 		contacts, err = db.QueryContacts()
 		if err != nil {
-			logrus.Info("no contacts")
-			return
-		}
-
-	}
-
-	output, err := json.Marshal(&contacts)
-	if err != nil {
-		logrus.Fatal(err)
-	}
-
-	if err := json.Indent(&result, output, "", " "); err != nil {
-		logrus.Fatal(err)
-	}
-
-	logrus.Info(string(result.Bytes()))
-
-	dir, err := cmd.Flags().GetString("output")
-	if err != nil {
-		logrus.Warn(err)
-		return
-	}
-
-	force, _ := cmd.Flags().GetBool("force")
-
-	if dir != "" {
-
-		// Create a file with the dirs.
-		if _, err := os.Stat(dir); err == nil && !force {
-			logrus.Warn("file exists, force flag required")
-		}
-
-		if err := ioutil.WriteFile(dir, result.Bytes(), 0644); err != nil {
 			logrus.Fatal(err)
 		}
-
 	}
+
+	app := &tui.App{
+		tview.NewApplication(),
+	}
+
+	table := app.Table(contacts)
+
+	if err := app.SetRoot(table, true).SetFocus(table).Run(); err != nil {
+		logrus.Fatal(err)
+	}
+
+	// dir, err := cmd.Flags().GetString("output")
+	// if err != nil {
+	// 	logrus.Warn(err)
+	// 	return
+	// }
+
+	// force, _ := cmd.Flags().GetBool("force")
+
+	// if dir != "" {
+
+	// 	// Create a file with the dirs.
+	// 	if _, err := os.Stat(dir); err == nil && !force {
+	// 		logrus.Warn("file exists, force flag required")
+	// 	}
+
+	// 	if err := ioutil.WriteFile(dir, result.Bytes(), 0644); err != nil {
+	// 		logrus.Fatal(err)
+	// 	}
+
+	// }
 
 }